Added setDefaultSettings button to all map edit bars except heat map
Added foolproofing for coordinates and zoom levels. Values that are out of range snap back to the closest valid value. Also added a Set Defaults Here button, that sets the default settings to the current center of the mapbox map. NOTE users still have to click save to actually set the values.